requirements.txt
streamlit==1.37.1
altair==5.2.0
streamlit-echarts==0.4.0
plotly==5.23.0
snowflake-connector-python==3.12.0
snowflake-sqlalchemy
pandas==1.5.3
joblib==1.4.2
Jinja2==3.1.4
numpy==1.26.4
scikit-learn==1.5.1
tensorflow==2.17.0
keras==3.4.1
pip-audit==2.7.3
python-dotenv==1.0.1
pytest==8.3.2

